@@ -10,17 +10,17 @@
         }
     ],
     "require": {
-        "kematjaya/export": "^2.0",
-        "symfony/dependency-injection": "^4.0|^5.0",
-        "symfony/config": "^4.0|^5.0",
-        "symfony/twig-bundle": "^4.0|^5.0",
-        "symfony/framework-bundle": "^4.0|^5.0",
-        "symfony/yaml": "^4.0|^5.0"
+        "kematjaya/export": "^7.0",
+        "symfony/dependency-injection": "^7.0",
+        "symfony/config": "^7.0",
+        "symfony/twig-bundle": "^7.0",
+        "symfony/framework-bundle": "^7.0",
+        "symfony/yaml": "^7.0"
     },
     "require-dev": {
-        "symfony/var-dumper": "^4.0|^5.0",
+        "symfony/var-dumper": "^7.0",
         "phpunit/phpunit": "^9.0",
-        "symfony/browser-kit": "^4.0|^5.0"
+        "symfony/browser-kit": "^7.0"
     },
     "autoload": {
         "psr-4": {
